Ported the example tests to pytest.
Instead of supplying bespoke tooling, instead rely on pytest for:

- test fixture generation
- diff generation
- timing
- logging and capture

In addition, offers features such as

- test selection / deselection
- unified invocation
