+++ tox.ini
@@ -9,7 +9,6 @@ setenv =
 
 commands =
   pytest {posargs}
-  python -m tests.test
 
 extras =
   test
